Use `test-helpers.js` for splitting between running on node and in browser.
