--- test/chai.js
@@ -1,2 +1,0 @@
-// use this file to split when run in the browser vs. run on node
-export { assert } from "chai";
