+++ test/tests/CalendarDayNamesHeader.tests.js
@@ -1,4 +1,4 @@
-import { assert } from '../chai.js';
+import { assert } from '../test-helpers.js';
 import { trimMarks } from "../normalize.js";
 import * as internal from "../../src/internal.js";
 import CalendarDayNamesHeader from "../../define/CalendarDayNamesHeader.js";
