+++ test/tests/FormElementMixin.tests.js
@@ -1,4 +1,4 @@
-import { assert } from '../chai.js';
+import { assert } from '../test-helpers.js';
 import * as internal from "../../src/internal.js";
 import FormElementMixin from "../../src/FormElementMixin.js";
 import ReactiveElement from "../../src/ReactiveElement.js";
