+++ test/tests/ShadowTemplateMixin.tests.js
@@ -1,4 +1,4 @@
-import { assert } from '../chai.js';
+import { assert } from '../test-helpers.js';
 import * as internal from "../../src/internal.js";
 import * as template from "../../src/template.js";
 import ShadowTemplateMixin from "../../src/ShadowTemplateMixin.js";
